Refactor grocery list loading and item addition
The changes improve the process of loading grocery items from a server and adding new items to the list. HTTP requests have been optimized, and the parsed data is more efficiently loaded to the grocery list. Furthermore, item addition has been altered from reloading the entire list to updating it with the newly saved item.