Handle API errors in grocery list widget
Implemented error handling for API requests in the Grocery List widget. Status codes over 400 or 'null' responses now trigger appropriate error messages to be displayed to the user. Code also added to manage the improper deletion of an item, by reinserting the item if an error occurs during the delete request.